@@ -15,7 +15,7 @@
     "build": "tsc",
     "lint": "eslint src dev --ext ts,tsx",
     "prepack": "yarn build",
-    "storybook": "sb dev",
+    "storybook": "storybook dev",
     "test": "jest --colors"
   },
   "files": [
@@ -25,21 +25,21 @@
     "CHANGELOG.md"
   ],
   "peerDependencies": {
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0"
+    "react": ">=18.3.1",
+    "react-dom": ">=18.3.1"
   },
   "dependencies": {
     "@gr4vy/embed": "^2.34.3",
     "lodash.isequal": "^4.5.0"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^7.6.20",
-    "@storybook/addons": "^7.6.20",
-    "@storybook/cli": "^7.6.20",
-    "@storybook/react": "^7.6.20",
-    "@storybook/react-webpack5": "^7.6.20",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@storybook/addon-essentials": "8.6.14",
+    "@storybook/addon-webpack5-compiler-babel": "^3.0.6",
+    "@storybook/cli": "8.6.14",
+    "@storybook/react": "8.6.14",
+    "@storybook/react-webpack5": "8.6.14",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "gitHead": "6606f868e978fc1aa7368557d2730d52170a9049"
 }
